backend/requirements.txt
fastapi
uvicorn
pymongo
pinecone-client
google-generativeai
python-dotenv